Ambiguous-Search-of-Documents/pyproject.toml
[tool.poetry]
name = "Ambiguous-Search-of-Documents"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
notebook = "^6.5.3"
numpy = "^1.25.2"
gensim = "^4.3.1"
mecab-python3 = "^1.0.6"
neologdn = "^0.5.2"
tqdm = "^4.66.1"
fasttext = "^0.9.2"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.10.3", extras = ["lint"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10.3"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py39"

[[tool.pysen.lint.mypy_targets]]
paths = ["."]
